re structuring folders, and Added mongodb Atlas
