+++ project webprog/views/pages/admin/products/create.ejs
@@ -42,11 +42,11 @@
             <input type="text" value="<%= product.sku %>" class="form-control" id="sku" name="sku" placeholder="">
           </div>
           <div class="form-group">
-            <label for="price">Harga produk : </label>
+            <label for="price">Harga produk (Rp) : </label>
             <input type="number" value="<%= product.price %>" class="form-control" id="price" name="price" placeholder="">
           </div>
           <div class="form-group">
-            <label for="weight">Berat produk : </label>
+            <label for="weight">Berat produk (Gram): </label>
             <input type="number" value="<%= product.weight %>" class="form-control" id="weight" name="weight" placeholder="">
           </div>
           <div class="form-group">
