+++ project webprog/views/pages/admin/products/detail.ejs
@@ -13,15 +13,23 @@
   <div class="content">
     <h1 style="font-family: 'Poppins';" class="text-center">Detail produk <%= product.nama %></h1>
     <div class="container">
-      <h2 class="text-muted">Thumbnail Produk :<img src="data:image/<%=product.img.contentType%>;base64, 
+      <h2 class="text-muted">Thumbnail Produk : <br><img src="data:image/<%=product.img.contentType%>;base64, 
         <%=product.img.data.toString('base64')%> " style="max-width: 200px;"> </h2>
+        <hr>
       <h2 class="text-muted">Produk SKU : <%= product.sku %></h2>
-      <h2 class="text-muted">Harga Produk : <%= product.price %></h2>
+      <hr>
+      <h2 class="text-muted">Harga Produk : Rp.<%= product.price %></h2>
+      <hr>
       <h2 class="text-muted">Berat Produk : <%= product.weight %> (gram)</h2>
+      <hr>
       <h2 class="text-muted">Deskripsi Produk : <%- product.desc %></h2>
+      <hr>
       <h2 class="text-muted">Merk : <%= product.brand %></h2>
+      <hr>
       <h2 class="text-muted">Kateogeri Produk : <%= product.category %></h2>
+      <hr>
       <h2 class="text-muted">Tanggal Input Produk : <%= product.createdAt.toLocaleDateString() %></h2>
+      <hr>
       <a href="/dashboard/product" class="btn btn-secondary">Semua produk</a>
       <a href="/dashboard/product/<%= product.id %>/edit" class="btn btn-danger">Edit</a>
     </div>
