+++ project webprog/views/pages/admin/products/index.ejs
@@ -41,7 +41,6 @@
                   <td class="table-data text-center"><%= product.sku %></td>
                   <td class="table-data text-center">Rp.<%= product.price %></td>
                   <td class="table-data  text-center"><%= product.weight %> (gram)</td>
-                  <td class="table-data text-center"><%- product.desc %></td>
                   <td class="table-data text-center"><%= product.brand %></td>
                   <td class="table-data text-center"><%= product.category %></td>
                   <td class="table-data text-center"><%= product.createdAt.toLocaleDateString() %></td>
