+++ project webprog/views/pages/index.ejs
@@ -81,7 +81,7 @@
 
         <div class="row">
           <div class="col-md-4 col-sm-12">
-            <a href="product/mouse">
+            <a href="/product/?category=Mouse">
               <div class="card-homepage mt-2 text-center ">
                 <img src="images/mouse-card-bg.jpg" class="card-homepage-bg" width="350px" height="100px" alt="">
                 <div class="card-homepage-title long"> <span class="underline--magical">Mouse</span></div>
@@ -91,7 +91,7 @@
 
           <div class="col-md-4 col-sm-12">
 
-            <a href="/product/headset">
+            <a href="/product/?category=Headset">
               <div class="card-homepage mt-2 text-center ">
                 <img src="images/headset-card-bg.jpg" class="card-homepage-bg" width="350px" height="100px" alt="">
                 <div class="card-homepage-title long"> <span class="underline--magical">Headset </span></div>
@@ -100,7 +100,7 @@
           </div>
 
           <div class="col-md-4 col-sm-12">
-            <a href="/product/keyboard">
+            <a href="/product/?category=Keyboard">
               <div class="card-homepage mt-2 text-center ">
                 <img src="images/keyboard-card-bg.jpg" class="card-homepage-bg" width="350px" height="100px" alt="">
                 <div class="card-homepage-title"> <span class="underline--magical">Keyboard </span></div>
@@ -110,7 +110,7 @@
         </div>
         <div class="row">
           <div class="col-md-4 col-sm-12">
-            <a href="/product/gaming-chair">
+            <a href="/product/?category=Gaming-c">
               <div class="card-homepage mt-2 text-center ">
                 <img src="images/chair-card-bg.jpg" class="card-homepage-bg" width="350px" height="100px" alt="">
                 <div class="card-homepage-title long"> <span class="underline--magical">Gaming Chair</span></div>
